requirements.txt
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
Django==5.2.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyaes==1.6.1
pyasn1==0.6.1
python-decouple==3.8
python-telegram-bot==22.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
Telethon==1.42.0
typing_extensions==4.15.0
openpyxl==3.1.5
Django>=4.2.0
openpyxl>=3.1.0
Telethon>=1.36.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
cryptg>=0.4.0
pytz>=2023.3